Rename rule, update severity, add more tests
