+++ tests/atest/rules/naming/mixed-task-test-settings/golden_test.robot
