+++ tests/atest/rules/naming/mixed-task-test-settings/mixed_tasks.robot
@@ -1,0 +1,12 @@
+*** Settings ***
+Task Setup      Some Keyword
+Test Teardown   Some Keyword
+Task Template   Some Keyword
+Test Timeout    100
+Task Tags       tag
+Documentation   some docs
+
+
+*** Tasks ***
+Some Task
+    No Operation
