+++ tests/atest/rules/naming/mixed-task-test-settings/tasks_with_test_settings.robot
