+++ tests/atest/rules/naming/mixed-task-test-settings/test_rule.py
@@ -1,0 +1,31 @@
+from tests.atest.utils import RuleAcceptance
+
+
+class TestRuleAcceptance(RuleAcceptance):
+    def test_rule_rf6(self):
+        self.check_rule(
+            src_files=[
+                "tasks_with_test_settings.robot",
+                "tests_with_task_settings.robot",
+                "mixed_tasks.robot",
+                "mixed_tests.robot",
+                "golden_task.robot",
+                "golden_test.robot",
+            ],
+            expected_file="expected_output_rf6.txt",
+            target_version="==6.*",
+        )
+
+    def test_rule(self):
+        self.check_rule(
+            src_files=[
+                "tasks_with_test_settings.robot",
+                "tests_with_task_settings.robot",
+                "mixed_tasks.robot",
+                "mixed_tests.robot",
+                "golden_task.robot",
+                "golden_test.robot",
+            ],
+            expected_file="expected_output_pre_rf6.txt",
+            target_version="<6",
+        )
