+++ tests/atest/rules/naming/mixed-task-test-settings/tests_with_task_settings.robot
